refactor: remove `promise-synchronizer` package
promise-synchronizer relies on the deasync library, which needs to be compiled with python 2. If the user does not have python 2 on the user's computer, an installation error will occur, and if the customer uses node14+, an installation error will occur, because node14+ uses python 3.

promise-synchronizer 依赖了 deasync 库, 这个库需要使用 python 2 编译. 如果使用者的电脑上没有 python 2 则安装出错, 如果客户使用 node14+ 则安装出错, 因为 node14+ 使用了 python 3.